deleated root element for CardView, some changes of app appearance, changed app's colors

+++ myapplication2/src/main/java/com/example/myapplication2/Utils.java
@@ -7,8 +7,8 @@ import com.bumptech.glide.load.engine.Resource;
 
 public class Utils {
 
-    public static int getWidth() {
-        return Resources.getSystem().getDisplayMetrics().widthPixels;
+    public static float getWidth() {
+        return (float)Resources.getSystem().getDisplayMetrics().widthPixels;
     }
 
     public static float getDpi() {
